Move the AssistAPI class into the llm integration
The AssistAPI class (and its Assist-specific prompt constants) now live in the
llm integration instead of homeassistant.helpers.llm, which keeps only the
framework (API base classes, Tool, IntentTool, the registry and the shared
selector serializer / exposed-entities helpers). The helper no longer references
AssistAPI at all.
